Revert to fpga-zynq aligning repo, tests cleanup
This reverts all of xfiles-dana to match a version of rocket-chip
that can be interfaced with fpga-zynq/master and, assumedly, the
Hurricane branch of rocket-chip. Reversions include:
  * Switching from `config` to `cde`
  * ANTW increments the addr on a put block for TileLink(1)

The unnecessary dependency on `config` in perfect-chisel is removed.

The tests are refactored to point directly at the
environment (riscv/riscv-env) directory included in a top-level
rocket-chip as opposed to using their own. This removes the need for
keeping these consistent.

The tests are refactored to put common Makefile code in a top-level
Makefrag and to follow standard Makefile naming conventions.

+++ src/main/scala/xfiles/XFiles.scala
@@ -6,7 +6,7 @@ package xfiles
 import chisel3._
 import chisel3.util._
 import rocket.{RoCC, HasCoreParameters, CoreModule, CoreBundle}
-import config._
+import cde._
 import math.pow
 import _root_.util.ParameterizedBundle
 import perfect.util.UniformPrintfs
@@ -62,7 +62,7 @@ trait XFilesResponseCodes extends HasCoreParameters with XFilesParameters
   val respCodeWidth = 3
 
   val (resp_OK :: resp_TID :: resp_READ :: resp_NOT_DONE :: resp_QUEUE_ERR ::
-    resp_XFILES :: Nil) =  Enum(UInt(), 6)
+    resp_XFILES :: Nil) =  Enum(6)
 
   def genResp[T <: Bits](resp: T, respCode: T, tid: T, data: T = 0.U(xLen.W)) {
     val tmp = Wire(new Bundle {
